refactor: [M3-5457] - convert to named export
