+++ packages/manager/src/components/BackupStatus/BackupStatus.tsx
@@ -6,7 +6,7 @@ import { Theme } from '@mui/material/styles';
 import Tooltip from 'src/components/core/Tooltip';
 import Typography from 'src/components/core/Typography';
 import DateTimeDisplay from 'src/components/DateTimeDisplay';
-import TooltipIcon from 'src/components/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
 import Link from 'src/components/Link';
 
 const useStyles = makeStyles((theme: Theme) => ({
