+++ packages/manager/src/components/Toggle/Toggle.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import ToggleOff from 'src/assets/icons/toggleOff.svg';
 import ToggleOn from 'src/assets/icons/toggleOn.svg';
 import Switch, { SwitchProps } from 'src/components/core/Switch';
-import TooltipIcon from 'src/components/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
 import './toggle.css';
 
 interface Props extends SwitchProps {
