+++ packages/manager/src/features/Billing/BillingPanels/BillingSummary/BillingSummary.tsx
@@ -6,7 +6,7 @@ import Divider from 'src/components/core/Divider';
 import Grid from '@mui/material/Unstable_Grid2';
 import PaymentDrawer from './PaymentDrawer';
 import PromoDialog from './PromoDialog';
-import TooltipIcon from 'src/components/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
 import Typography from 'src/components/core/Typography';
 import useAccountManagement from 'src/hooks/useAccountManagement';
 import { ActivePromotion } from '@linode/api-v4/lib/account/types';
