+++ packages/manager/src/features/Users/UserProfile.tsx
@@ -11,7 +11,7 @@ import Paper from 'src/components/core/Paper';
 import { Theme, useTheme } from '@mui/material/styles';
 import Typography from 'src/components/core/Typography';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
-import TooltipIcon from 'src/components/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
 import Notice from 'src/components/Notice';
 import TextField from 'src/components/TextField';
 import { useProfile } from 'src/queries/profile';
