+++ packages/manager/src/features/linodes/LinodesCreate/SelectPlanPanel.tsx
@@ -15,7 +15,7 @@ import TableBody from 'src/components/core/TableBody';
 import TableHead from 'src/components/core/TableHead';
 import Typography from 'src/components/core/Typography';
 import Currency from 'src/components/Currency';
-import TooltipIcon from 'src/components/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
 import Grid from '@mui/material/Unstable_Grid2';
 import Notice from 'src/components/Notice';
 import Radio from 'src/components/Radio';
