+++ packages/manager/src/features/linodes/LinodesLanding/LinodeRow/LinodeRow.tsx
@@ -12,7 +12,7 @@ import Flag from 'src/assets/icons/flag.svg';
 import Hidden from 'src/components/core/Hidden';
 import Tooltip from 'src/components/core/Tooltip';
 import Typography from 'src/components/core/Typography';
-import TooltipIcon from 'src/components/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
 import StatusIcon from 'src/components/StatusIcon';
 import TableCell from 'src/components/TableCell';
 import TableRow from 'src/components/TableRow';
